task1-wine-recommendation/pyproject.toml
[tool.poetry]
name = "vinobuzz-sommelier"
version = "0.1.0"
description = "VinoBuzz AI Sommelier — Multi-agent wine recommendation system"
authors = ["VinoBuzz <it@vino-intel.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
openai-agents = "^0.0.17"
openai = "^1.50.0"
sse-starlette = "^2.0.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
